requirements.txt
google-genai
python-dotenv
streamlit
nest_asyncio